two changes for making progress on aspnet/DataCommon.SQLite#21
add support for sqlite3_db_filename.

also, add sqlite3__vfs__delete().  instead of publicly exposing
sqlite3_vfs_find(), which would require non-portable marshaling
code to be used to call the function pointers within the struct,
I added a function which calls sqlite3_vfs_find() and does the
marshaling and then calls xDelete().  other calls in the sqlite3_vfs
struct could be handled the same way later.  also later, I could
expose a real sqlite3_vfs object, but for now, since the DataCommon.SQLite
project seems to only be using this as a way of deleting a file,
this seems sufficient.

the doubled underscores in the name of sqlite3__vfs__delete() are
intentional, since this function is not in the low-level sqlite C
API.  I want it to look a bit different and to avoid clashes with
future functions.

no progress in this commit on the AnyCPU-at-build-time issues.