Downgrade AnyCPU Error to Warning
In the scenario where the target is a platform-specific library (i.e. a
library whose TargetFramework is Net45), it is entirely valid to link to
SQLitePCL.Raw but still be AnyCPU.

This restriction should really only pop up if the target is an App, but
defining that on every platform is way trickier (i.e. a unit test DLL is
technically an "app" in the sense that it isn't a library being integrated as
a piece of another program, and therefore should be x86 or else the tests will
fail).