Experimental change. Move SQLiteProvider_esqlite3 and SQLiteProvider_sqlcipher into the main assembly.
And then, the initialization code (in the static constructor for raw)
tries to load those other providers before settling for the system SQLite.

This has been mentioned and considered before.  I previously disliked the idea
because it won't work for iOS with __Internal.  But with the Android N SQLite
change, I'm starting to think of iOS as the oddball case.

SetProvider(ISQLiteProvider) still exists.  It is used for iOS,
and can be used for custom provider implementations.

I also added a built-in provider for "custom_sqlite3", so if you want to use
a custom SQLite build, just give it that name and it should be found automatically.

I also added an enum of the names of those built-in providers, and a SetProvider(enum)
overload which can be used to override which provider got chosen by the static constructor.

Having multiple built-in providers means there is a search order to define
which one is preferred if more than one are present.
Right now, that ordering is:

    custom_sqlite3 (never included by me)
    sqlcipher
    esqlite3 (special name for sqlite libs included by me)
    sqlite3 (system sqlite)

One possible future change would be to remove the system sqlite3
from this "search path" on Android builds.

Another: might be to have iOS builds ONLY look for the system sqlite,
since the DllImport calls are useless there, unless we start
providing sqlite on iOS as shared libraries, which is another
can of worms.

Another: might be to add the name of the system SQLite on
Windows 10 (winsqlite3, I believe).

So, in terms of the code itself, I think this would cause no breakage.
If the system SQLite is the only thing found, the behavior of the static constructor
init code should be unchanged.  Any calls to SetProvider will still work.

However, this change motivates a couple of possible changes to packaging.

Currently, I generate 10 nuget packages which are "empty plugins".  They contain
only the Provider implementation with no native code.  But in all of those cases,
that Provider implementation has moved into the main assembly.  So now the package
is truly empty and doesn't need to exist.  So in this commit, those 10 nuget
packages are not generated anymore.

Now, if I eliminate those 10 packages, the naming of the remaining set looks a bit weird.
The iOS + Android packages are called "plugin", and the other platforms are called "native".
This leaves me a bit tempted to pick one name and eliminate the other one, as the different
names exist mostly because of the strange road I walked to get here.  OTOH, if we keep the
names unchanged, we could explain it by saying that the Xamarin-ish ones have a different
name because they bundle the native code inside an assembly as a resource, whereas the
other platforms bundle the native code as shared libraries inside nuget's directories.
For now, I am leaving the somewhat-dorky naming.

Bottom line:

The benefit of all this is that it preserves fine-grained control for people
who need it, while making more things Just Work for the common cases.  On iOS, anything
but the system SQLite still requires an initialization call.  Everywhere else, most
users can just add a nuget package and things will work.