Fix test_other_col_types when running in time zones with a positive offset from UTC
This test would only pass when run in time zones with a negative offset from UTC. Since the test only asserts on year, month and day, the test passes. When run in a timezone with a positive offset from UTC, the day would change. For example, in UTC+2, the date would become 1968-03-14T22:00:00.0000000 and the test would fail (14 != 15).

By doing `d.ToUniversalTime()` the time of the supplied DateTime was changed and the stored date was actually altered (by its offset from UTC). Asserting on the whole date would fail:
>    Assert.Equal() Failure  
> Expected: 1968-03-15T00:00:00.0000000  
> Actual:   1968-03-15T08:00:00.0000000  

The assertion has been changed to assert on the full DateTime and the code was fixed to not alter the supplied date. So now, `test_other_col_types` always passes, no matter what `TimeZoneInfo.Local` is.